refactor(web-embed): serve the embed-builder script same-origin, not inline
The embed builder page injected its copy-to-clipboard + live PAGE_URL
substitution behaviour as an inline <script>. The web skill ("Browser JS Is
Bundled and Served Same-Origin") forbids inline scripts (a CSP liability).

Move the IIFE into embed-client-script.ts and serve it at
/embed/embed.client.js (same pattern as the existing /embed/icon.svg route),
referenced from the page with <script src> instead of inline. web-embed has
no client-TS bundler, so the script stays a self-executing string served as a
static same-origin asset.

Audit finding: high priority — Browser JS must be bundled and served same-origin.
