+++ projects/web-embed/src/runtime/embed/embed.component.ts
@@ -9,48 +9,6 @@ import { byteLength, renderCanonicalSnippet, renderSnippet } from "./snippet.com
 
 const EMBED_TEMPLATE = readFileSync(join(__dirname, "embed.template.html"), "utf-8");
 
-const COPY_SCRIPT = `<script>
-(function() {
-  var buttons = document.querySelectorAll('[data-copy]');
-  for (var i = 0; i < buttons.length; i++) {
-    (function(btn) {
-      btn.addEventListener('click', function() {
-        var target = document.getElementById(btn.getAttribute('data-copy'));
-        if (!target) return;
-        navigator.clipboard.writeText(target.textContent || '').then(function() {
-          var original = btn.textContent;
-          btn.textContent = 'Copied';
-          setTimeout(function() { btn.textContent = original; }, 1500);
-        });
-      });
-    })(buttons[i]);
-  }
-
-  var snippetIds = ['snippet-a-code', 'snippet-b-code', 'snippet-c-code'];
-  var originals = {};
-  for (var i = 0; i < snippetIds.length; i++) {
-    var el = document.getElementById(snippetIds[i]);
-    if (el) originals[snippetIds[i]] = el.textContent;
-  }
-
-  var urlInput = document.querySelector('.embed-url-input__field');
-  if (urlInput) {
-    urlInput.addEventListener('input', function() {
-      var url = urlInput.value.trim();
-      for (var id in originals) {
-        var el = document.getElementById(id);
-        if (!el) continue;
-        if (url) {
-          el.textContent = originals[id].replace('PAGE_URL', encodeURIComponent(url));
-        } else {
-          el.textContent = originals[id];
-        }
-      }
-    });
-  }
-})();
-</script>`;
-
 const CANONICAL_EMBED_ORIGIN = "https://readplace.com/embed";
 
 export interface EmbedPageInput {
@@ -91,7 +49,7 @@ export function EmbedPage(input: EmbedPageInput): Component {
 		pageStyles: EMBED_PAGE_STYLES,
 		bodyClass: "page-embed",
 		content,
-		scripts: COPY_SCRIPT,
+		scripts: `<script src="${input.embedOrigin}/embed.client.js" defer></script>`,
 		appOrigin: input.appOrigin,
 	});
 }
